(limits): moved SkillChangeEvent
